Unifiying naming convention between nuget and normla http repo

+++ Bootstrap/Program.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Configuration;
+using System.Diagnostics;
 using System.IO;
 using System.IO.Compression;
 using System.Net;
@@ -13,21 +14,78 @@ namespace CSLauncher.Packager
         static int Main(string[] args)
         {
             string url = ConfigurationManager.AppSettings.Get("url");
-            string downloadLocation  = Path.Combine(ConfigurationManager.AppSettings.Get("location"), "Deployer");
-            if (!Directory.Exists(downloadLocation))
+            Version version = new Version(ConfigurationManager.AppSettings.Get("version"));
+            string deployerLocation = Environment.ExpandEnvironmentVariables(ConfigurationManager.AppSettings.Get("location"));
+            string deployerSentinel = Path.Combine(deployerLocation, "__deployer__");
+
+            bool download = false;
+            if (!Directory.Exists(deployerLocation))
             {
-                Directory.CreateDirectory(downloadLocation);
+                Directory.CreateDirectory(deployerLocation);
+                download = true;
             }
-            string zipFile = Path.Combine(downloadLocation, "Deployer.zip");
-            using (var client = new WebClient())
+            else
             {
-                client.DownloadFile(url, zipFile);
+                if (File.Exists(deployerSentinel))
+                {
+                    string sentinelContent = File.ReadAllText(deployerSentinel);
+                    Version sentinelVersion = new Version(sentinelContent);
+                    if (version > sentinelVersion)
+                    {
+                        DirectoryInfo dir = new DirectoryInfo(deployerLocation);
+                        FileInfo[] files = dir.GetFiles();
+                        foreach (FileInfo file in files)
+                        {
+                            file.Delete();
+                        }
+                        download = true;
+                    }
+                }
+                else
+                {
+                    download = true;
+                }
+            }
+            if (download)
+            {
+                string zipFile = Path.Combine(deployerLocation, "Deployer.zip");
+                using (var client = new WebClient())
+                {
+                    client.DownloadFile(url, zipFile);
+                }
+
+                ZipFile.ExtractToDirectory(zipFile, deployerLocation);
+                File.Delete(zipFile);
+                File.WriteAllText(deployerSentinel, version.ToString());
             }
 
-            ZipFile.ExtractToDirectory(zipFile, downloadLocation);
-            File.Delete(zipFile);
+            string deployerExe = Path.Combine(deployerLocation, "Deployer.exe");
+            ProcessStartInfo startInfo = new ProcessStartInfo();
+            startInfo.UseShellExecute = false;
+            startInfo.WorkingDirectory = Environment.CurrentDirectory;
+            startInfo.FileName = deployerExe;
+            string cmd = Environment.CommandLine;
+            // Remove exe part from raw command line
+            int firstArgIndex = 0;
+            bool discardNextSpaces = false;
+            foreach (char c in cmd)
+            {
+                if (c == '\"')
+                {
+                    discardNextSpaces = !discardNextSpaces;
+                }
 
-            return 0;
+                if (!discardNextSpaces && (c == ' ' || c == '\t'))
+                    break;
+
+                firstArgIndex++;
+            }
+            string arguments = cmd.Substring(firstArgIndex, cmd.Length - firstArgIndex).Trim();
+            startInfo.Arguments = arguments;
+
+            Process p = Process.Start(startInfo);
+            p.WaitForExit();
+            return p.ExitCode;
         }
     }
 }
